Refactor API integration
