+++ src/containers/ListUsers.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import Title from '../components/Title';
 import Button from '../components/Button';
 
-import { getUsers } from '../data/UserFetcher';
+import { getUsers } from '../data/user';
 
 import './ListUsers.css';
